Extract into ReportingAbstractRequest class
