+++ src/Message/FetchTransactionsRequest.php
@@ -2,7 +2,6 @@
 
 namespace Omnipay\BlueSnap\Message;
 
-use Omnipay\BlueSnap\Constants;
 use Omnipay\Common\Exception\InvalidRequestException;
 
 /**
@@ -47,52 +46,7 @@ use Omnipay\Common\Exception\InvalidRequestException;
  *   }
  * </code>
  */
-class FetchTransactionsRequest extends AbstractRequest
+class FetchTransactionsRequest extends ReportingAbstractRequest
 {
     const REPORT_NAME = 'TransactionDetail';
-
-    /**
-     * @return null
-     */
-    public function getData()
-    {
-        $this->validate('startTime');
-        $this->validate('endTime');
-
-        if ($this->getStartTime() > $this->getEndTime()) {
-            throw new InvalidRequestException('startTime cannot be greater than endTime');
-        }
-
-        return null;
-    }
-
-    /**
-     * Return the API endpoint.
-     *
-     * @return string
-     */
-    public function getEndpoint()
-    {
-        $startTime = $this->getStartTime();
-        $endTime = $this->getEndTime();
-        $endpoint = parent::getEndpoint() . '/report/' . (string) static::REPORT_NAME;
-
-        // this should always be true
-        if ($startTime && $endTime) {
-            $endpoint .= '?period=CUSTOM'
-                      . '&from_date=' . urlencode((string) $startTime->format('m/d/Y'))
-                      . '&to_date=' . urlencode((string) $endTime->format('m/d/Y'));
-        }
-        return $endpoint;
-    }
-
-    /**
-     * Returns the HTTP method to be used for this request
-     *
-     * @return string
-     */
-    public function getHttpMethod()
-    {
-        return Constants::HTTP_METHOD_GET;
-    }
 }
